refactor: HomeController를 헤더(새그룹추가) 컴포넌트로 분리하고 HomeForm 재구성 (#15)
