+++ src/app/routes/layout.tsx
@@ -1,15 +1,19 @@
 import EventFulLogo from '@/shared/ui/eventFulLogo';
-import { Outlet } from 'react-router-dom';
+import { Outlet, useNavigate } from 'react-router-dom';
 import Sidebar from './sidebar';
 
 interface LayoutProps {
   isSidebar: boolean;
 }
 export default function Layout({ isSidebar }: LayoutProps) {
+  const navigate = useNavigate();
+
   return (
     <div>
       <div className="flex justify-between h-[50px] border-b border-gray-300 px-6">
-        <EventFulLogo />
+        <div className="cursor-pointer" onClick={() => navigate('/home')}>
+          <EventFulLogo />
+        </div>
         <div className="y-center space-x-3">
           <span className="text-sm font-medium bg-[#009689] text-white px-4 py-1 rounded-[8px]">
             낙현
